MA-6899 [Android_MX6QP/SX_ARD]Brightness: NAND boot is darker than SDCard boot for 6QP, and brightness is dark for 6SX_ARD. 100%
1 The brightness on imx6sx_ard with nand\sd boot is darker than other board.
The hannstar star cabc is enable when boot. The root cause is the pin LVDS_EN is connected at i2c expander a I\O7,
the configure for cabc pin is not correct.

2 The brightness with nand boot is darker than sd boot on imx6q_ard\imx6qp_ard.
The i2c3 is disable for it's pin is confict with weim when boot up with nand. So cabc driver is not loaded when nand boot up,
and did not set BACKLITE_ON to low.

Signed-off-by: zhang sanshan <b51434@freescale.com>